-added better logging for worker rush defense, and production queue -smarter requeue of CC vs cannon rush -no bunker rush vs sharkybot -updated launchers with new maps -unblocking refinery if sieged tank is in the way -cancelling structures now makes the structure position available again in the future -cannon rush defense now ignores probes when worker rush defense is also active

+++ src/main/java/com/ketroc/utils/Time.java
@@ -55,13 +55,13 @@ public class Time {
     }
 
     //returns true if within first numFrames of any numMinutes period
-    public static boolean periodic(int numMinutes, int numFrames) {
-        long periodFrame = nowFrames() % (NUM_FRAMES_PER_MINUTE * numMinutes);
+    public static boolean periodic(float numMinutes, int numFrames) {
+        float periodFrame = nowFrames() % (NUM_FRAMES_PER_MINUTE * numMinutes);
         return periodFrame < numFrames;
     }
 
     //returns true if within first played frame of any numMinutes period
-    public static boolean periodic(int numMinutes) {
+    public static boolean periodic(float numMinutes) {
         return periodic(numMinutes, Launcher.STEP_SIZE);
     }
